Add optional target layer info to shapes move

+++ server/src/api/socket/shape/__init__.py
@@ -137,11 +137,17 @@ async def send_remove_shapes(
     await _send_game("Shapes.Remove", data, room=room, skip_sid=skip_sid)
 
 
-def _get_shapes_from_uuids(uuids: list[str]) -> SelectSequence[Shape]:
-    return Shape.select().where(
+def _get_shapes_from_uuids(
+    uuids: list[str], filter_layer: bool
+) -> SelectSequence[Shape]:
+    query = Shape.select().where(
         (Shape.uuid << uuids)  # pyright: ignore[reportGeneralTypeIssues]
-        & ~(Shape.layer >> None)  # pyright: ignore[reportGeneralTypeIssues]
     )
+    if filter_layer:
+        query = query.where(
+            ~(Shape.layer >> None)  # pyright: ignore[reportGeneralTypeIssues]
+        )
+    return query
 
 
 @sio.on("Shapes.Remove", namespace=GAME_NS)
@@ -157,7 +163,7 @@ async def remove_shapes(sid: str, raw_data: Any):
             game_state.remove_temp(sid, shape)
     else:
         # Use the server version of the shapes.
-        shapes = list(_get_shapes_from_uuids(data.uuids))
+        shapes = list(_get_shapes_from_uuids(data.uuids, True))
         if len(shapes) == 0:
             return
 
@@ -177,11 +183,15 @@ async def remove_shapes(sid: str, raw_data: Any):
             if shape.group:
                 group_ids.add(shape.group)
 
-            old_index = shape.index
-            shape.delete_instance(True)
-            Shape.update(index=Shape.index - 1).where(
-                (Shape.layer == layer) & (Shape.index >= old_index)
-            ).execute()
+            if shape.character_id is None:
+                old_index = shape.index
+                shape.delete_instance(True)
+                Shape.update(index=Shape.index - 1).where(
+                    (Shape.layer == layer) & (Shape.index >= old_index)
+                ).execute()
+            else:
+                shape.layer = None
+                shape.save()
 
         for group_id in group_ids:
             await remove_group_if_empty(group_id)
@@ -202,7 +212,7 @@ async def change_shape_floor(sid: str, raw_data: Any):
         return
 
     floor: Floor = Floor.get(location=pr.active_location, name=data.floor)
-    shapes = list(_get_shapes_from_uuids(data.uuids))
+    shapes = list(_get_shapes_from_uuids(data.uuids, True))
     layer: Layer = Layer.get(
         floor=floor,
         name=shapes[0].layer.name,  # pyright: ignore[reportOptionalMemberAccess]
@@ -236,7 +246,7 @@ async def change_shape_layer(sid: str, raw_data: Any):
         return
 
     floor = Floor.get(location=pr.active_location, name=data.floor)
-    shapes = list(_get_shapes_from_uuids(data.uuids))
+    shapes = list(_get_shapes_from_uuids(data.uuids, True))
     layer = Layer.get(floor=floor, name=data.layer)
     old_layer = shapes[0].layer
 
@@ -355,21 +365,26 @@ async def move_shapes(sid: str, raw_data: Any):
     location = Location.get_by_id(data.target.location)
     floor = location.floors.where(Floor.name == data.target.floor)[0]
 
-    shapes = list(_get_shapes_from_uuids(data.shapes))
-    layer = shapes[0].layer
-    if layer is None:
-        logger.error("Attempt to location-move shape without layer")
-        return
+    target_layer = None
+    if data.target.layer:
+        target_layer = floor.layers.where(Layer.name == data.target.layer)[0]
+
+    shapes = []
+    for shape in _get_shapes_from_uuids(data.shapes, False):
+        layer = target_layer
+        if shape.layer:
+            layer = floor.layers.where(Layer.name == shape.layer.name)[0]
+        elif layer is None:
+            logger.warn("Attempt to move a shape without layer info")
+            continue
+        shapes.append((shape, layer))
 
     await send_remove_shapes(
-        [sh.uuid for sh in shapes], room=layer.floor.location.get_path()
+        [sh.uuid for sh, _ in shapes], room=floor.location.get_path()
     )
 
-    for shape in shapes:
-        if shape.layer is None:
-            continue
-
-        shape.layer = floor.layers.where(Layer.name == shape.layer.name)[0]
+    for shape, layer in shapes:
+        shape.layer = layer
         shape.center = data.target
         shape.save()
 
@@ -382,7 +397,7 @@ async def move_shapes(sid: str, raw_data: Any):
                     floor=floor.name,
                     layer=layer.name,
                 )
-                for shape in shapes
+                for shape, layer in shapes
             ],
             room=psid,
         )
